+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to the Audibene Code Snippets will be documented here.
 - Snippets for Theme Cards
 - Updated Classes
 - Updated Classes in Advanced Mode 
-- Included if and if else in Aura
+- Included `if`, `if empty`, `if not empty`, `if equals`, `if not equals`, `if else`, `else` in Aura
 
 ## Amber - v1.0.2 
 - 10 New Snippets added for JavaScript